pii_service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6

# Presidio (Analyzer+Anonymizer) + SpaCy NLP engine
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.7.5
phonenumbers==8.13.45

# GLiNER zero-shot NER (local HF model)
gliner==0.2.10
torch>=2.1.0

# pydantic v1 compatible with FastAPI 0.115
pydantic==1.10.15
python-dotenv==1.0.1
